+++ CHANGELOG.rst
@@ -7,8 +7,9 @@ Release 0.5
 
 - Added block for metadata
 - New icons from Font Awesome (http://fortawesome.github.io/Font-Awesome/icons/)
-- New pagination option: Dropdown
-- New context processor, **site_path**, to display site and site root urls
+- New dropdown pagination template
+- New context processor, **site_path**, for access to the current configured
+  site via django.contrib.sites
 
 Release 0.4
 -----------